MILESTONE: Add initial performance metrics
The network runner script outputs a CSV with the ping times for every
node to every other node, done twice per timestep. Now that the TTL of
the openflow rules have been increased to 120s, and the links changed to
have non-negligible latency, This means that the first set of ping
metrics at each times step (batch zero), mainly tests the controller's
ability to install routes in the switches (if lazy routing is done), and
the second set of metrics assesses optimality.